+++ src/contract/contractMethod.js
@@ -2,30 +2,29 @@
  * @file contract method
  * @author atom-yang
  */
-import {
-  getTransaction,
-  Transaction
-} from '../util/proto';
+import { getTransaction, Transaction } from '../util/proto';
 import {
   transformArrayToMap,
   transformMapToArray,
   transform,
   INPUT_TRANSFORMERS,
-  OUTPUT_TRANSFORMERS
+  OUTPUT_TRANSFORMERS,
 } from '../util/transform';
 import {
   isBoolean,
   isFunction,
+  isNumber,
   noop,
   uint8ArrayToHex,
-  unpackSpecifiedTypeData
+  unpackSpecifiedTypeData,
 } from '../util/utils';
 import wallet from '../wallet';
 
 export default class ContractMethod {
-  constructor(chain, method, contractAddress, walletInstance) {
+  constructor(chain, method, contractAddress, walletInstance, option) {
     this._chain = chain;
     this._method = method;
+    this._option = option || {};
     const { resolvedRequestType, resolvedResponseType } = method;
     this._inputType = resolvedRequestType;
     this._outputType = resolvedResponseType;
@@ -61,7 +60,7 @@ export default class ContractMethod {
     }
     const result = unpackSpecifiedTypeData({
       data: inputPacked,
-      dataType: this._inputType
+      dataType: this._inputType,
     });
     let params = transform(this._inputType, result, OUTPUT_TRANSFORMERS);
     params = transformArrayToMap(this._inputType, params);
@@ -73,8 +72,8 @@ export default class ContractMethod {
       return null;
     }
     let result = unpackSpecifiedTypeData({
-      data: output,
-      dataType: this._outputType
+      data: isNumber(output) ? String(output) : output,
+      dataType: this._outputType,
     });
     result = transform(this._outputType, result, OUTPUT_TRANSFORMERS);
     result = transformArrayToMap(this._outputType, result);
@@ -103,29 +102,88 @@ export default class ContractMethod {
     return uint8ArrayToHex(tx);
   }
 
-  prepareParametersAsync(args) {
-    const filterArgs = args.filter(arg => !isFunction(arg) && !isBoolean(arg.sync));
+  prepareParametersAsync(args, isView) {
+    const filterArgs = args.filter(
+      arg => !isFunction(arg) && !isBoolean(arg.sync)
+    );
     const encoded = this.packInput(filterArgs[0]);
 
+    if (isView) {
+      return Promise.resolve(this.handleTransaction('', '', encoded));
+    }
     return this._chain.getChainStatus().then(status => {
-      const { BestChainHeight, BestChainHash } = status;
+      let { BestChainHeight, BestChainHash } = status;
+
+      let { refBlockNumberStrategy } = this._option || {};
+
+      args.forEach(arg => {
+        if (arg.refBlockNumberStrategy) {
+          // eslint-disable-next-line max-len
+          if (typeof arg.refBlockNumberStrategy !== 'number') throw new Error('Invalid type, refBlockNumberStrategy must be number');
+          if (arg.refBlockNumberStrategy > 0) throw new Error('refBlockNumberStrategy must be less than 0');
+          refBlockNumberStrategy = arg.refBlockNumberStrategy;
+        }
+      });
+
+      if (refBlockNumberStrategy) {
+        BestChainHeight += refBlockNumberStrategy;
+        const block = this._chain.getBlockByHeight(BestChainHeight, true, {
+          sync: true
+        });
+        BestChainHash = block.BlockHash;
+      }
+
       return this.handleTransaction(BestChainHeight, BestChainHash, encoded);
     });
   }
 
-  prepareParameters(args) {
-    const filterArgs = args.filter(arg => !isFunction(arg) && !isBoolean(arg.sync));
+  /**
+   * @param {Array} args - argument
+   * @param {boolean} isView - view method
+   * @returns any
+   */
+  prepareParameters(args, isView) {
+    const filterArgs = args.filter(
+      arg => !isFunction(arg) && !isBoolean(arg.sync)
+    );
     const encoded = this.packInput(filterArgs[0]);
 
-    const { BestChainHeight, BestChainHash } = this._chain.getChainStatus({
-      sync: true
+    if (isView) {
+      return this.handleTransaction('', '', encoded);
+    }
+
+    let { refBlockNumberStrategy } = this._option;
+
+    args.forEach(arg => {
+      if (arg.refBlockNumberStrategy) {
+        // eslint-disable-next-line max-len
+        if (typeof arg.refBlockNumberStrategy !== 'number') throw new Error('Invalid type, refBlockNumberStrategy must be number');
+        if (arg.refBlockNumberStrategy > 0) throw new Error('refBlockNumberStrategy must be less than 0');
+        refBlockNumberStrategy = arg.refBlockNumberStrategy;
+      }
     });
+
+    const statusRes = this._chain.getChainStatus({
+      sync: true,
+    });
+
+    let { BestChainHeight, BestChainHash } = statusRes;
+
+    if (refBlockNumberStrategy) {
+      BestChainHeight += refBlockNumberStrategy;
+      const block = this._chain.getBlockByHeight(BestChainHeight, true, {
+        sync: true
+      });
+      BestChainHash = block.BlockHash;
+    }
 
     return this.handleTransaction(BestChainHeight, BestChainHash, encoded);
   }
 
   prepareParametersWithBlockInfo(args) {
-    const filterArgs = args.filter(arg => !isFunction(arg) && !isBoolean(arg.sync));
+    const filterArgs = args.filter(
+      arg => !isFunction(arg) && !isBoolean(arg.sync)
+    );
     const encoded = this.packInput(filterArgs[0]);
 
     const { height, hash } = filterArgs[1]; // blockInfo
@@ -138,7 +196,7 @@ export default class ContractMethod {
     if (argsObject.isSync) {
       const parameters = this.prepareParameters(args);
       return this._chain.sendTransaction(parameters, {
-        sync: true
+        sync: true,
       });
     }
     // eslint-disable-next-line arrow-body-style
@@ -150,23 +208,27 @@ export default class ContractMethod {
   callReadOnly(...args) {
     const argsObject = this.extractArgumentsIntoObject(args);
     if (argsObject.isSync) {
-      const parameters = this.prepareParameters(args);
-      return this.unpackOutput(this._chain.callReadOnly(parameters, {
-        sync: true
-      }));
+      const parameters = this.prepareParameters(args, true);
+      return this.unpackOutput(
+        this._chain.callReadOnly(parameters, {
+          sync: true,
+        })
+      );
     }
     // eslint-disable-next-line arrow-body-style
-    return this.prepareParametersAsync(args).then(parameters => {
-      return this._chain.callReadOnly(parameters, (error, result) => {
-        argsObject.callback(error, this.unpackOutput(result));
-      }).then(this.unpackOutput);
+    return this.prepareParametersAsync(args, true).then(parameters => {
+      return this._chain
+        .callReadOnly(parameters, (error, result) => {
+          argsObject.callback(error, this.unpackOutput(result));
+        })
+        .then(this.unpackOutput);
     });
   }
 
   extractArgumentsIntoObject(args) {
     const result = {
       callback: noop,
-      isSync: false
+      isSync: false,
     };
     if (args.length === 0) {
       // has no callback, default to be async mode
@@ -176,7 +238,7 @@ export default class ContractMethod {
       result.callback = args[args.length - 1];
     }
     args.forEach(arg => {
-      if (isBoolean((arg.sync))) {
+      if (isBoolean(arg.sync)) {
         result.isSync = arg.sync;
       }
     });
@@ -185,7 +247,9 @@ export default class ContractMethod {
 
   // getData(...args) {
   getSignedTx(...args) {
-    const filterArgs = args.filter(arg => !isFunction(arg) && !isBoolean(arg.sync));
+    const filterArgs = args.filter(
+      arg => !isFunction(arg) && !isBoolean(arg.sync)
+    );
 
     if (filterArgs[1]) {
       const { height, hash } = filterArgs[1]; // blockInfo
@@ -199,11 +263,21 @@ export default class ContractMethod {
   }
 
   getRawTx(blockHeightInput, blockHashInput, packedInput) {
-    const rawTx = getTransaction(this._wallet.address, this._contractAddress, this._name, packedInput);
-
-    rawTx.refBlockNumber = blockHeightInput;
-    const blockHash = blockHashInput.match(/^0x/) ? blockHashInput.substring(2) : blockHashInput;
-    rawTx.refBlockPrefix = (Buffer.from(blockHash, 'hex')).slice(0, 4);
+    const rawTx = getTransaction(
+      this._wallet.address,
+      this._contractAddress,
+      this._name,
+      packedInput
+    );
+    if (blockHeightInput) {
+      rawTx.refBlockNumber = blockHeightInput;
+    }
+    if (blockHashInput) {
+      const blockHash = blockHashInput.match(/^0x/)
+        ? blockHashInput.substring(2)
+        : blockHashInput;
+      rawTx.refBlockPrefix = Buffer.from(blockHash, 'hex').slice(0, 4);
+    }
     return rawTx;
   }
 
@@ -214,7 +288,7 @@ export default class ContractMethod {
       method: 'broadcast_tx',
       callback,
       params,
-      format: this.unpackOutput
+      format: this.unpackOutput,
     };
   }