+++ src/index.js
@@ -3,7 +3,6 @@
  * @author atom-yang
  */
 import * as protobuf from '@aelfqueen/protobufjs/light';
-import jsSha256 from 'js-sha256';
 import * as bloom from './util/bloom';
 import Chain from './chain';
 import RequestManager from './util/requestManage';
@@ -13,8 +12,8 @@ import * as utils from './util/utils';
 import * as proto from './util/proto';
 import * as transform from './util/transform';
 import Settings from './util/settings';
+import sha256 from './util/sha256';
 
-const { sha256 } = jsSha256;
 /* eslint-disable no-underscore-dangle */
 export default class AElf {
   constructor(provider) {